requirements.txt
langchain==1.2.10
langchain-core==1.2.12
langchain-openai==1.1.9
openai==2.21.0
pydantic==2.12.5
python-dotenv==1.2.1
streamlit==1.54.0
tiktoken==0.12.0